+++ packages/api-shared/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @doccov/api-shared
 
+## 0.2.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @doccov/sdk@0.25.2
+
 ## 0.2.5
 
 ### Patch Changes