+++ packages/sandbox/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @doccov/sandbox
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @doccov/sdk@0.25.2
+
 ## 0.1.2
 
 ### Patch Changes